Refactor source code away from lightcone.h
This is to prepare for the removal.
